--- include/VectorAlgebra.hpp
@@ -1,27 +1,0 @@
-#pragma once
-#include <cassert>
-#include <iostream>
-#include <string>
-#include <type_traits>
-#include <stdlib.h>
-#include <utility>
-#include <cstdlib>
-
-#include<vectorAlgebra/vector.hpp>
-#include<vectorAlgebra/vectorUnit.hpp>
-
-// Extra<helper stuff to keep things clean
-#include<vectorAlgebra/templateHelpers.hpp>
-#include<vectorAlgebra/MAX.hpp>
-#include<vectorAlgebra/SUM.hpp>
-
-#include<vectorAlgebra/elementOperations.hpp>
-#include<vectorAlgebra/vectorPairwiseMap.hpp>
-#include<vectorAlgebra/vectorMap.hpp>
-
-#include<vectorAlgebra/operators.hpp>
-#include<vectorAlgebra/operations.hpp>
-
-#include<vectorAlgebra/infNorm.hpp>
-
-#include<vectorAlgebra/traits.hpp>
