+++ include/panoc/VectorAlgebra.hpp
@@ -1,0 +1,26 @@
+#pragma once
+#include <cassert>
+#include <iostream>
+#include <string>
+#include <type_traits>
+#include <stdlib.h>
+#include <utility>
+#include <cstdlib>
+
+#include<panoc/vectorAlgebra/vector.hpp>
+#include<panoc/vectorAlgebra/vectorUnit.hpp>
+
+#include<panoc/vectorAlgebra/templateHelpers.hpp>
+#include<panoc/vectorAlgebra/MAX.hpp>
+#include<panoc/vectorAlgebra/SUM.hpp>
+
+#include<panoc/vectorAlgebra/elementOperations.hpp>
+#include<panoc/vectorAlgebra/vectorPairwiseMap.hpp>
+#include<panoc/vectorAlgebra/vectorMap.hpp>
+
+#include<panoc/vectorAlgebra/operators.hpp>
+#include<panoc/vectorAlgebra/operations.hpp>
+
+#include<panoc/vectorAlgebra/infNorm.hpp>
+
+#include<panoc/vectorAlgebra/traits.hpp>
