+++ include/panoc/vectorAlgebra/MAX.hpp
