+++ include/panoc/vectorAlgebra/elementOperations.hpp
