+++ include/panoc/vectorAlgebra/operators.hpp
