+++ include/panoc/vectorAlgebra/templateHelpers.hpp
