+++ include/panoc/vectorAlgebra/vectorMap.hpp
